requirements/base.txt
Django==5.1.*
djangorestframework==3.15.*
djangorestframework-simplejwt==5.3.*
drf-nested-routers==0.93.*
django-filter==24.*
drf-spectacular==0.27.*
psycopg[binary]==3.*
django-environ==0.11.*
gunicorn==22.*
whitenoise==6.*
django-cors-headers==4.*
Pillow==11.*
